Updated stop_container.sh

+++ scripts/stop_container.sh
@@ -2,5 +2,10 @@
 set -e
 
 # Stop the running container (if any)
-containerid = `docker ps | awk -F " " '{print $1}'`
-docker container rm -f $containerid
+containerid=$(docker ps -q)
+
+if [ -n "$containerid" ]; then
+  docker container rm -f $containerid
+else
+  echo "No running container to stop."
+fi
